moved ado_time to library.
