+++ ado_chip_175x_6x/src/ado_time.c
@@ -5,7 +5,7 @@
  *      Author: Robert
  */
 
-#include "ado_time.h"
+#include <ado_time.h>
 #include <math.h>
 
 juliandayfraction TimeGetCurrentDayOfYear();
